# Implemented overflow detection in the current parseInt implementation.