**Description**: What this PR does
Fixes Issue #138 - Adds int64 overflow check in the current `parseInt` implementation.

I used the benchmarks in `bytes_test` to check performance:
```
 go test -test.benchmem -bench ParseInt ./ -benchtime 5s -run ^A$ -v
goos: darwin
goarch: amd64
pkg: github.com/jstroem/jsonparser
BenchmarkParseInt-8               	1000000000	         7.93 ns/op	       0 B/op	       0 allocs/op
BenchmarkParseIntUnsafeSlower-8   	500000000	        14.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkParseIntOverflows-8      	2000000000	         5.96 ns/op	       0 B/op	       0 allocs/op
```
Note that `BenchmarkParseIntOverflows` is the old implementation. This shows that the fix costs `~2ns/op`.

**Benchmark before change**:
```
goos: darwin
goarch: amd64
pkg: github.com/jstroem/jsonparser/benchmark
BenchmarkJsonParserLarge-8                    	  200000	     55469 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-8                   	 1000000	      8912 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserDeleteMedium-8             	 1000000	      7331 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-8      	 1000000	      5181 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-8      	 1000000	      5782 ns/op	     672 B/op	      12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-8   	 1000000	      8676 ns/op	     624 B/op	      11 allocs/op
BenchmarkJsonParserSmall-8                    	10000000	       831 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-8       	10000000	       618 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-8       	10000000	       706 ns/op	     144 B/op	       4 allocs/op
BenchmarkJsonParserObjectEachStructSmall-8    	10000000	       626 ns/op	     128 B/op	       3 allocs/op
BenchmarkJsonParserSetSmall-8                 	10000000	      1146 ns/op	     816 B/op	       5 allocs/op
BenchmarkJsonParserDelSmall-8                 	 5000000	      1379 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/jstroem/jsonparser/benchmark	99.531s
``` 

**Benchmark after change**:
```
goos: darwin
goarch: amd64
pkg: github.com/jstroem/jsonparser/benchmark
BenchmarkJsonParserLarge-8                    	  200000	     56137 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-8                   	 1000000	      8997 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserDeleteMedium-8             	 1000000	      7398 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-8      	 1000000	      5180 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-8      	 1000000	      5810 ns/op	     672 B/op	      12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-8   	 1000000	      8891 ns/op	     624 B/op	      11 allocs/op
BenchmarkJsonParserSmall-8                    	10000000	       878 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-8       	10000000	       665 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-8       	10000000	       739 ns/op	     144 B/op	       4 allocs/op
BenchmarkJsonParserObjectEachStructSmall-8    	10000000	       648 ns/op	     128 B/op	       3 allocs/op
BenchmarkJsonParserSetSmall-8                 	10000000	      1200 ns/op	     816 B/op	       5 allocs/op
BenchmarkJsonParserDelSmall-8                 	 5000000	      1597 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/jstroem/jsonparser/benchmark	103.480s
```